test: reerrange tests
